renaming firmware to Electron
