+++ src/vdp/Makefile
@@ -9,7 +9,7 @@ OBJS = $(SRCS:.cpp=.o)
 .cpp.o:
 	$(CXX) $(CFLAGS) $(OPTIONS) -fPIC -c $< -o $@
 
-all: userspace-vdp-gl/vdp_gl.a vdp_console8.so vdp_quark103.so vdp_quark104.so vdp_electronhal.so test_main
+all: userspace-vdp-gl/vdp_gl.a vdp_console8.so vdp_quark103.so vdp_quark104.so vdp_electron.so test_main
 
 userspace-vdp-gl/vdp_gl.a:
 	$(MAKE) -C userspace-vdp-gl/src
@@ -26,15 +26,15 @@ vdp_quark103.so: $(OBJS) userspace-vdp-gl/vdp_gl.a vdp-1.03.o
 vdp_quark104.so: $(OBJS) userspace-vdp-gl/vdp_gl.a vdp_quark104.o
 	$(CXX) -shared $(OBJS) vdp_quark104.o userspace-vdp-gl/src/vdp-gl.a -o vdp_quark104.so
 
-vdp_electronhal.so: $(OBJS) userspace-vdp-gl/vdp_gl.a vdp_electronhal.o
-	$(CXX) -shared $(OBJS) vdp_electronhal.o userspace-vdp-gl/src/vdp-gl.a -o vdp_electronhal.so
+vdp_electron.so: $(OBJS) userspace-vdp-gl/vdp_gl.a vdp_electron.o
+	$(CXX) -shared $(OBJS) vdp_electron.o userspace-vdp-gl/src/vdp-gl.a -o vdp_electron.so
 
 clean:
 	-rm *.o
 	-rm *.so
 	$(MAKE) -C userspace-vdp-gl/src clean
 
-SRC_ALL = $(SRCS) vdp-1.03.cpp vdp-console8.cpp vdp_quark104.cpp vdp_electronhal.cpp
+SRC_ALL = $(SRCS) vdp-1.03.cpp vdp-console8.cpp vdp_quark104.cpp vdp_electron.cpp
 
 depends: $(SRC_ALL)
 	$(CXX) -MM $(CFLAGS) $(SRC_ALL) > Makefile.dep
